Move BindingAlias to separate file
